Added gun script commit
+ gun script
- player shooting script
