+++ Assets/Scripts/Gun.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 16fb910428ea96f4c9280146a1fec152
+guid: bcd9a27c1e8ecf14f99a29a93f51b98a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
